feat: Add BFS class for performing breadth-first search
The BFS class extends the Graph class and includes methods for performing a breadth-first traversal of a graph, counting the number of connected components in the graph, and computing the transitive closure of the graph. The class initializes the level array to Integer.MAX_VALUE, initializes and executes the breadth-first traversal of the graph starting at a given vertex, and counts the number of connected components in the graph. Additionally, the class computes the transitive closure of the graph and returns a boolean matrix M such that M[i][j] is true if there is a path from i to j.